+++ lib/signalwire/relay/calling/action/play_action.rb
@@ -11,5 +11,17 @@ module Signalwire::Relay::Calling
     def stop
       @component.stop
     end
+
+    def pause
+      @component.pause
+    end
+
+    def resume
+      @component.resume
+    end
+
+    def volume(setting)
+      @component.volume setting
+    end
   end
 end