+++ lib/signalwire/relay/calling/call_convenience_methods.rb
@@ -1,11 +1,19 @@
 module Signalwire::Relay::Calling
   module CallConvenienceMethods
-    def play_audio(url)
-      play audio_payload(url)
+    def play_audio(url_p = nil, volume_p = nil, **args)
+      url = args.delete(:url)
+      volume = args.delete(:volume)
+      set_parameters(binding, %i{url volume}, %i{url})
+
+      play(play: audio_payload(url), volume: volume)
     end
 
-    def play_audio!(url)
-      play! audio_payload(url)
+    def play_audio!(url_p = nil, volume_p = nil, **args)
+      url = args.delete(:url)
+      volume = args.delete(:volume)
+      set_parameters(binding, %i{url volume}, %i{url})
+
+      play!(play: audio_payload(url), volume: volume)
     end
 
     def play_silence(duration)
@@ -16,33 +24,35 @@ module Signalwire::Relay::Calling
       play! silence_payload(duration)
     end
 
-    def play_tts(text_p=nil, language_p=Relay::DEFAULT_LANGUAGE, gender_p=Relay::DEFAULT_GENDER, text: nil, language: Relay::DEFAULT_LANGUAGE, gender: Relay::DEFAULT_GENDER)
-      set_parameters(binding, %i{text language gender}, %i{text})
+    def play_tts(text_p=nil, language_p=Relay::DEFAULT_LANGUAGE, gender_p=Relay::DEFAULT_GENDER, volume_p=nil, text: nil, language: Relay::DEFAULT_LANGUAGE, gender: Relay::DEFAULT_GENDER, volume: nil)
+      set_parameters(binding, %i{text language gender volume}, %i{text})
 
-      play tts_payload(text, language, gender)
+      play play: tts_payload(text, language, gender), volume: volume
     end
 
-    def play_tts!(text_p=nil, language_p=Relay::DEFAULT_LANGUAGE, gender_p=Relay::DEFAULT_GENDER, text: nil, language: Relay::DEFAULT_LANGUAGE, gender: Relay::DEFAULT_GENDER)
-      set_parameters(binding, %i{text language gender}, %i{text})
-      play! tts_payload(text, language, gender)
+    def play_tts!(text_p=nil, language_p=Relay::DEFAULT_LANGUAGE, gender_p=Relay::DEFAULT_GENDER, text: nil, language: Relay::DEFAULT_LANGUAGE, gender: Relay::DEFAULT_GENDER, volume: nil)
+      set_parameters(binding, %i{text language gender volume}, %i{text})
+      play! tts_payload(text, language, gender), volume: volume
     end
 
-    def prompt_audio(collect_p = nil, url_p = nil, **args)
+    def prompt_audio(collect_p = nil, url_p = nil, volume_p=nil, **args)
       collect = args.delete(:collect)
       url = args.delete(:url)
+      volume = args.delete(:volume)
       collect = compile_collect_arguments(args) if collect.nil? && collect_p.nil?
-      set_parameters(binding, %i{collect url}, %i{collect url})
+      set_parameters(binding, %i{collect url volume}, %i{collect url})
 
-      prompt(collect: collect, play: audio_payload(url))
+      prompt(collect: collect, play: audio_payload(url), volume: volume)
     end
 
-    def prompt_audio!(collect_p = nil, url_p = nil, **args)
+    def prompt_audio!(collect_p = nil, url_p = nil, volume_p=nil, **args)
       collect = args.delete(:collect)
       url = args.delete(:url)
+      volume = args.delete(:volume)
       collect = compile_collect_arguments(args) if collect.nil? && collect_p.nil?
-      set_parameters(binding, %i{collect url}, %i{collect url})
+      set_parameters(binding, %i{collect url volume}, %i{collect url})
 
-      prompt!(collect: collect, play: audio_payload(url))
+      prompt!(collect: collect, play: audio_payload(url), volume: volume)
     end
 
     def prompt_silence(collect_p = nil, duration_p = nil, **args)
@@ -63,28 +73,30 @@ module Signalwire::Relay::Calling
       prompt!(collect: collect, play: silence_payload(duration))
     end
 
-    def prompt_tts(collect_p = nil, text_p = nil, language_p=Relay::DEFAULT_LANGUAGE, gender_p=Relay::DEFAULT_GENDER, **args)
+    def prompt_tts(collect_p = nil, text_p = nil, volume_p=nil, language_p=Relay::DEFAULT_LANGUAGE, gender_p=Relay::DEFAULT_GENDER, **args)
       collect = args.delete(:collect)
       text = args.delete(:text)
+      volume = args.delete(:volume)
       language = args.delete(:language) || Relay::DEFAULT_LANGUAGE
       gender = args.delete(:gender) || Relay::DEFAULT_GENDER
       collect = compile_collect_arguments(args) if collect.nil? && collect_p.nil?
 
-      set_parameters(binding, %i{collect text language gender}, %i{collect text})
+      set_parameters(binding, %i{collect text language gender volume}, %i{collect text})
 
-      prompt(collect: collect, play: tts_payload(text, language, gender))
+      prompt(collect: collect, play: tts_payload(text, language, gender), volume: volume)
     end
 
-    def prompt_tts!(collect_p = nil, text_p = nil, language_p=Relay::DEFAULT_LANGUAGE, gender_p=Relay::DEFAULT_GENDER, **args)
+    def prompt_tts!(collect_p = nil, text_p = nil, volume_p=nil, language_p=Relay::DEFAULT_LANGUAGE, gender_p=Relay::DEFAULT_GENDER, **args)
       collect = args.delete(:collect)
       text = args.delete(:text)
+      volume = args.delete(:volume)
       language = args.delete(:language) || Relay::DEFAULT_LANGUAGE
       gender = args.delete(:gender) || Relay::DEFAULT_GENDER
       collect = compile_collect_arguments(args) if collect.nil? && collect_p.nil?
 
-      set_parameters(binding, %i{collect text language gender}, %i{collect text})
+      set_parameters(binding, %i{collect text language gender volume}, %i{collect text})
 
-      prompt!(collect: collect, play: tts_payload(text, language, gender))
+      prompt!(collect: collect, play: tts_payload(text, language, gender), volume: volume)
     end
 
     def wait_for_ringing