+++ lib/signalwire/relay/calling/control_component.rb
@@ -24,11 +24,15 @@ module Signalwire::Relay::Calling
     end
 
     def stop
-      @call.relay_execute execute_params('.stop') do |event, outcome|
+      execute_subcommand '.stop', Signalwire::Relay::Calling::StopResult
+    end
+
+    def execute_subcommand(suffix, result_klass, extra_params = {})
+      @call.relay_execute execute_params('.stop', extra_params) do |event, outcome|
         succeeded = outcome == :success
         terminate(event) unless succeeded
 
-        return Signalwire::Relay::Calling::StopResult.new(succeeded)
+        return result_klass.new(succeeded)
       end
     end
   end