+++ lib/signalwire/relay/calling/result/base_action_result.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+module Signalwire::Relay::Calling
+  class BaseActionResult
+    attr_reader :successful
+    def initialize(result)
+      @successful = result
+    end
+  end
+end