+++ lib/signalwire/relay/calling/result/play_resume_result.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+module Signalwire::Relay::Calling
+  class PlayResumeResult < BaseActionResult
+  end
+end