+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add `pause` and `resume` on `PlayAction`.
+- Add `volume` optional parameter to `play` and `prompt` methods
+- Add `volume` method to `Play` and `Prompt` components
 
 ## [2.2.0] - 2019-09-09
 ### Added